@@ -352,10 +336,9 @@
     <!--参数设置-->
     <properties>
         <spring.version>4.3.9.RELEASE</spring.version>
-        <java.version>1.7</java.version>
+        <java.version>1.8</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <jetty.version>8.1.9.v20130131</jetty.version>
         <port>8888</port>
-        <jackson.version>1.9.12</jackson.version>
     </properties>
 </project>
