@@ -22,17 +22,18 @@ import java.util.Date;
     private static Logger logger = LoggerFactory.getLogger(FileServiceImpl.class);
     @Autowired private FileMapper fileMapper;
 
-    @Override public int addFile(String fname, int comid, String flocal, String fremarks) {
+    @Override public int addFile(String fname, int pid, String flocal, String fsummary) {
         int fid = 1;
         File file = new File();
         file.setFname(fname);
-        //        file.setComid(comid);
-        //        file.setFremarks(fremarks);
+                file.setPid(pid);
+                file.setFsummary(fsummary);
         //TODO
-        file.setFlocal("/file/contractfile/" + fname);
+        file.setFlocal("src/main/webapp/uploadFiles/" + fname);
         Date date = DateUtil.getNowDate();
         file.setCreatetime(date);
         file.setModtime(date);
+        file.setUid(1);
         //        file.setEmpid(this.comempRepository.findByUid(uid).getEmpid());
         //        file.setBrid(this.comempRepository.findByUid(uid).getBid());
         try {
@@ -45,7 +46,7 @@ import java.util.Date;
         return fid;
     }
 
-    @Override public String editFile(String fname, int comid, String flocal, String fremarks) {
+    @Override public String editFile(String fname, int comid, String flocal, String fsummary) {
         //TODO
         //      Subject subject = SecurityUtils.getSubject();
         //      Session session = subject.getSession();
@@ -60,6 +61,7 @@ import java.util.Date;
         Date date = DateUtil.getNowDate();
         file.setModtime(date);
         file.setCreatetime(new Date());
+        file.setUid(1);
         //        file.setEmpid(this.comempRepository.findByUid(uid).getEmpid());
         //        file.setBrid(this.comempRepository.findByUid(uid).getBid());
         try {
